requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==40.0.2
cssselect==1.2.0
dataclasses-json==0.5.7
dnspython==2.3.0
filelock==3.12.0
Flask==2.3.2
frozenlist==1.3.3
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.6.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
langchain==0.0.172
loguru==0.7.0
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
parsel==1.8.1
pinecone-client==2.2.1
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
PyDispatcher==2.0.7
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
queuelib==1.6.2
regex==2023.5.5
requests==2.30.0
requests-file==1.5.1
Scrapy==2.9.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
tenacity==8.2.2
tiktoken==0.4.0
tldextract==3.4.4
tqdm==4.65.0
Twisted==22.10.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==2.0.2
w3lib==2.1.1
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
